Stabilize production planning logic around a single scheduling path
This change consolidates the active production planning path around the
new package layout, restores fractional BOM quantity support, and removes
duplicated scheduler logic that was no longer part of the runtime entry
surface. It also records the next logical refactor steps in a durable plan.

Constraint: The repository contains broad unrelated changes, so this commit stages only the production_planning logic subset
Constraint: Matching and feasibility behavior had to remain compatible with the active scheduler and existing tests
Rejected: Keep the legacy parallel scheduler modules alongside engine.py | dead path increased maintenance surface and diverging logic risk
Rejected: Partial float support only in nomenclature | would leave hidden integer truncation in downstream logic
Confidence: medium
Scope-risk: moderate
Reversibility: clean
Directive: Treat apps/ordo-core/production_planning as the canonical production planning runtime path; do not reintroduce duplicate scheduler stacks without an explicit migration plan
Tested: pytest in apps/ordo-core (189 passed)
Not-tested: Live ERP/API flows and board-ui integration against these logic changes

+++ packages/erp-data-access/src/erp_data_access/models/gamme.py
@@ -32,13 +32,13 @@ class Gamme:
                 return op
         return None
 
-    def calculate_hours(self, quantity: int, poste: str) -> float:
+    def calculate_hours(self, quantity: float, poste: str) -> float:
         op = self.get_operation_for_poste(poste)
         if not op or op.cadence == 0:
             return 0.0
         return quantity / op.cadence
 
-    def calculate_all_hours(self, quantity: int) -> dict[str, float]:
+    def calculate_all_hours(self, quantity: float) -> dict[str, float]:
         result = {}
         for op in self.operations:
             if op.cadence > 0:
